// apps/trackflix/src/API.ts
export enum Source {
  SELF = "SELF",
  YOUTUBE = "YOUTUBE",
  LIVESTREAM_SELF = "LIVESTREAM_SELF",
}
// ...
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export enum ModelAttributeTypes {
  binary = "binary",
  binarySet = "binarySet",
  bool = "bool",
  list = "list",
  map = "map",
  number = "number",
  numberSet = "numberSet",
  string = "string",
  stringSet = "stringSet",
  _null = "_null",
}


export type ModelSizeInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
};

export type ModelBooleanInput = {
  ne?: boolean | null,
  eq?: boolean | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type ModelSourceInput = {
  eq?: Source | null,
  ne?: Source | null,
};

export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type Media = {
  __typename: "Media",
  id: string,
  title: string,
  description: string,
  highlighted: boolean,
  source?: Source | null,
  author: string,
  viewCount?: number | null,
  createdAt: string,
  updatedAt: string,
  thumbnail?: Thumbnail | null,
  sections?: ModelMediasSectionsConnection | null,
};

export type Thumbnail = {
  __typename: "Thumbnail",
  id: string,
  ext: string,
  src?: string | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelMediasSectionsConnection = {
  __typename: "ModelMediasSectionsConnection",
  items:  Array<MediasSections | null >,
  nextToken?: string | null,
};

export type MediasSections = {
  __typename: "MediasSections",
  id: string,
  sectionID: string,
  mediaID: string,
  createdAt: string,
  updatedAt: string,
  media: Media,
  section: Section,
};

export type Section = {
  __typename: "Section",
  id: string,
  label: string,
  description: string,
  createdAt: string,
  updatedAt: string,
  medias?: ModelMediasSectionsConnection | null,
};
// ...
export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type VideoOnDemand = {
  __typename: "VideoOnDemand",
  id: string,
  metadata?: string | null,
  src?: string | null,
  createdAt: string,
  updatedAt: string,
  media?: Media | null,
};
// ...
export type SecuredLink = {
  __typename: "SecuredLink",
  id: string,
  mediaID: string,
  usagesRemaining: number,
  notifiedEmail?: string | null,
  expirationUnixTime: number,
  createdAt: string,
  updatedAt: string,
  media: Media,
};
// ...
export type ModelMediaConnection = {
  __typename: "ModelMediaConnection",
  items:  Array<Media | null >,
  nextToken?: string | null,
};
// ...
export type ModelSectionConnection = {
  __typename: "ModelSectionConnection",
  items:  Array<Section | null >,
  nextToken?: string | null,
};
// ...
export type ModelSecuredLinkFilterInput = {
  id?: ModelIDInput | null,
  mediaID?: ModelIDInput | null,
  usagesRemaining?: ModelIntInput | null,
  notifiedEmail?: ModelStringInput | null,
  expirationUnixTime?: ModelIntInput | null,
  and?: Array< ModelSecuredLinkFilterInput | null > | null,
  or?: Array< ModelSecuredLinkFilterInput | null > | null,
  not?: ModelSecuredLinkFilterInput | null,
};

export type ModelSecuredLinkConnection = {
  __typename: "ModelSecuredLinkConnection",
  items:  Array<SecuredLink | null >,
  nextToken?: string | null,
};

export type ModelSecuredLinkLogFilterInput = {
  id?: ModelIDInput | null,
  mediaID?: ModelIDInput | null,
  securedLinkId?: ModelIDInput | null,
  message?: ModelStringInput | null,
  and?: Array< ModelSecuredLinkLogFilterInput | null > | null,
  or?: Array< ModelSecuredLinkLogFilterInput | null > | null,
  not?: ModelSecuredLinkLogFilterInput | null,
};
// ...
export type ListMediaQuery = {
  listMedia?:  {
    __typename: "ModelMediaConnection",
    items:  Array< {
      __typename: "Media",
      id: string,
      title: string,
      description: string,
      highlighted: boolean,
      source?: Source | null,
      author: string,
      viewCount?: number | null,
      createdAt: string,
      updatedAt: string,
      thumbnail?:  {
        __typename: "Thumbnail",
        id: string,
        ext: string,
        src?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      sections?:  {
        __typename: "ModelMediasSectionsConnection",
        nextToken?: string | null,
      } | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};
// ...
export type GetSectionQuery = {
  getSection?:  {
    __typename: "Section",
    id: string,
    label: string,
    description: string,
    createdAt: string,
    updatedAt: string,
    medias?:  {
      __typename: "ModelMediasSectionsConnection",
      items:  Array< {
        __typename: "MediasSections",
        id: string,
        sectionID: string,
        mediaID: string,
        createdAt: string,
        updatedAt: string,
      } | null >,
      nextToken?: string | null,
    } | null,
  } | null,
};

export type ListSectionsQueryVariables = {
  filter?: ModelSectionFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListSectionsQuery = {
  listSections?:  {
    __typename: "ModelSectionConnection",
    items:  Array< {
      __typename: "Section",
      id: string,
      label: string,
      description: string,
      createdAt: string,
      updatedAt: string,
      medias?:  {
        __typename: "ModelMediasSectionsConnection",
        nextToken?: string | null,
      } | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetMediasSectionsQueryVariables = {
  id: string,
};

export type GetMediasSectionsQuery = {
  getMediasSections?:  {
    __typename: "MediasSections",
    id: string,
    sectionID: string,
    mediaID: string,
    createdAt: string,
    updatedAt: string,
    media:  {
      __typename: "Media",
      id: string,
      title: string,
      description: string,
      highlighted: boolean,
      source?: Source | null,
      author: string,
      viewCount?: number | null,
      createdAt: string,
      updatedAt: string,
      thumbnail?:  {
        __typename: "Thumbnail",
        id: string,
        ext: string,
        src?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      sections?:  {
        __typename: "ModelMediasSectionsConnection",
        nextToken?: string | null,
      } | null,
    },
    section:  {
      __typename: "Section",
      id: string,
      label: string,
      description: string,
      createdAt: string,
      updatedAt: string,
      medias?:  {
        __typename: "ModelMediasSectionsConnection",
        nextToken?: string | null,
      } | null,
    },
  } | null,
};

export type ListMediasSectionsQueryVariables = {
  filter?: ModelMediasSectionsFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListMediasSectionsQuery = {
  listMediasSections?:  {
    __typename: "ModelMediasSectionsConnection",
    items:  Array< {
      __typename: "MediasSections",
      id: string,
      sectionID: string,
      mediaID: string,
      createdAt: string,
      updatedAt: string,
      media:  {
        __typename: "Media",
        id: string,
        title: string,
        description: string,
        highlighted: boolean,
        source?: Source | null,
        author: string,
        viewCount?: number | null,
        createdAt: string,
        updatedAt: string,
      },
      section:  {
        __typename: "Section",
        id: string,
        label: string,
        description: string,
        createdAt: string,
        updatedAt: string,
      },
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetVideoOnDemandQueryVariables = {
  id: string,
};

export type GetVideoOnDemandQuery = {
  getVideoOnDemand?:  {
    __typename: "VideoOnDemand",
    id: string,
    metadata?: string | null,
    src?: string | null,
    createdAt: string,
    updatedAt: string,
    media?:  {
      __typename: "Media",
      id: string,
      title: string,
      description: string,
      highlighted: boolean,
      source?: Source | null,
      author: string,
      viewCount?: number | null,
      createdAt: string,
      updatedAt: string,
      thumbnail?:  {
        __typename: "Thumbnail",
        id: string,
        ext: string,
        src?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      sections?:  {
        __typename: "ModelMediasSectionsConnection",
        nextToken?: string | null,
      } | null,
    } | null,
  } | null,
};
// ...
export type GetSecuredLinkQuery = {
  getSecuredLink?:  {
    __typename: "SecuredLink",
    id: string,
    mediaID: string,
    usagesRemaining: number,
    notifiedEmail?: string | null,
    expirationUnixTime: number,
    createdAt: string,
    updatedAt: string,
    media:  {
      __typename: "Media",
      id: string,
      title: string,
      description: string,
      highlighted: boolean,
      source?: Source | null,
      author: string,
      viewCount?: number | null,
      createdAt: string,
      updatedAt: string,
      thumbnail?:  {
        __typename: "Thumbnail",
        id: string,
        ext: string,
        src?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      sections?:  {
        __typename: "ModelMediasSectionsConnection",
        nextToken?: string | null,
      } | null,
    },
  } | null,
};